Chore: unify sql data driver implementation
